Flush commands when changing hotkeys on X11
Apparently not every X11 connection flushes the commands automatically.
However, it's really important for the hotkey changes to take effect
immediately. We don't do anything else on the X11 connection, so the
changes would be pending until the next hotkey changes or possibly even
later.

This additionally cleans up the implementation to not ungrab all keys
every time a hotkey change happens. This should not affect the bug, but
it did not seem like the right thing to do in general.