Support the `web` Target of `wasm-bindgen` (#852)
This emits the bindings in a way such that they are compatible with not
only the `bundler` target of `wasm-bindgen`, but also the `web` target.
This allows us to circumvent `webpack`, which means we are not limited
by the features it supports anymore. In particular, this unblocks the
usage of reference types.