Fix Bounding Box Calculation (#859)
* Fix Bounding Box Calculation

At some point when changing our text engine, we introduced a scale value
for each individual glyph, possibly during the transition to
`cosmic-text`. This is correctly being used when rendering the text, but
not when calculating the bounding box of the text. This meant that for
certain fonts the dirty region of the image, that is cleared on each
frame, was not calculated correctly and thus the previous text was still
visible.

* Fix WASI 0.1 Target

It got renamed to mention the version number in the target name.