Move Text Shadow Support to Rendering Backends
Instead of duplicating every label, it is much better for the rendering
backends to handle the text shadow support by themselves. This also
enables them to do it in a faster or more sophisticated way, such as
using a proper text shadow blur.

Web Backend:

We use the `shadowColor`, `shadowOffsetX`, `shadowOffsetY`, and
`shadowBlur` properties of the `CanvasRenderingContext2D` to draw the
text shadows.

Software Backend:

We still basically just duplicate the labels. `tiny-skia` does not have
a good way to add shadows or have any support for blurring.

SVG Backend:

We unfortunately also just duplicate the labels. The reason is that we
tried to use a `feDropShadow` filter, but it seems to be cut off in
Chrome and weirdly enough **segfaults** Firefox.

https://x.com/CryZe107/status/1903132316358082813