Fix Generated TypeScript Types (#888)
Apparently `Uint8Array` is ambiguous in the latest TypeScript versions,
as it can be backed by either an `ArrayBuffer` or a `SharedArrayBuffer`
now. If you are working with an ambiguous `Uint8Array`, you can't
necessarily use it everywhere, such as when creating a `Blob` out of it.
We should be precise about which ones we are returning out of our C API
functions.

Additionally, this implements a slight improvement in the web renderer
around empty strings. Previously, there were various places where we had
to create an empty string. Now, we create it once and reuse it.