Add support for importing more than just the "Personal Best" comparison
The only function for importing a comparison, "import_comparison", would
only ever read the "Personal Best" comparison from the file when
importing. This commit adds the "import_comparison_as_comparison"
function, which allows for a named comparison from the file to be
specified, instead of just loading "Personal Best".

Due to this addition, "import_comparison" is reduced to calling
"import_comparison_as_comparison", with the name specified directly as
"Personal Best".